\name{daily.moral}
\alias{daily.moral}
\docType{data}
\title{An example data set}
\usage{
data(daily.moral)
}
\description{
This dataset is extracted from the dataset of a large study by Hardy et al. (2014) on the relationships among daily religious involvement, spirituality and moral emotions. The dataset contains 50 days of self-ratings of various moral emotions (empathy, gratitude and forgiveness) from a single participant.
}
\references{
Hardy, S. A., Zhang, Z., Skalski, J. E., & Melling, B. S. (2014) Daily religious involvement, spirituality, and moroal emotions. Psychology of Religion and Spirituality, 6, 338-348. doi: 10.1037/a0037293
}
\examples{
head(daily.moral)
}
