\name{Durbin_Levinson_rho}
\alias{Durbin_Levinson_rho}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Estimate Partial Autocorrelations Using the Durbin-Levinson Algorithm
}
\description{
This function computes partial autocorrelations using the Durbin-Levinson algorithm
}
\usage{Durbin_Levinson_rho(rho)}

\arguments{
  \item{rho}{a vector of autocorrelations}
}
\details{
}

\value{a vector of partial autocorrelations.}

\references{
Shumway, R. H., & Stoffer, D. S. (2006). Time series analysis and its applications with r examples (2nd ed.). New York: Springer.  (p.113-114)
}
\author{Zijun Ke <keziyun@mail.sysu.edu.cn> and Zhiyong Zhang <zhiyongzhang@nd.edu>}
\note{
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
