\name{get.ahat}
\alias{get.ahat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute the Acceleration
}
\description{
 compute the acceleration for bias-corrected and acceleration intervals
}
\usage{
get.ahat(dv)
}
\arguments{
  \item{dv}{
	vector of resampling estimates to compute the acceleration
}
}

\value{
	the accelaration estimate
}
\references{
Efron, B., & Tibshirani, R. J. (1994). An introduction to the bootstrap. Chapman and Hall/CRC.
}
\author{
Zijun Ke <keziyun@mail.sysu.edu.cn> and Zhiyong Zhang <zhiyongzhang@nd.edu>
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
