\name{pautocorr-package}
\alias{pautocorr-package}
\alias{pautocorr}
\docType{package}
\title{
Asymptotic methods and resampling techniques for testing (partial) autocorrelations
}
\description{
Two asymptotic methods (the 1/T approximation method and the Bartlett's formula method) and two resampling techniques (the surrogate data method and the vectorized moving block bootstrap) are implemented to test the significance of (partial) autocorrelations at certain lags.
}

\author{
Author: Zijun Ke and Zhiyong Zhang
Maintainer: Zijun Ke <keziyun@mail.sysu.edu.cn>
}

