\name{plot.pautocorr}
\alias{plot.pautocorr}
\alias{plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot the results
}
\description{
Plot the results}
\usage{
\method{plot}{pautocorr}(x, method,
	plot.control = list(layout.v = c(2,2),
	legendpos = list(A='bottomright',B ='topleft')),...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Results from the function \code{\link{pautocorr.test}}.
}
  \item{method}{
To add
}
  \item{plot.control}{
To add}
  \item{legendpos}{
To add}
\item{...}{Options can be passed to the \code{\link{plot}} function.}
}
