\name{Durbin_Levinson_rho}
\alias{Durbin_Levinson_rho}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Estimate Partial Autocorrelations Using the Durbin-Levinson Algorithm
}
\description{
This function computes partial autocorrelations using the Durbin-Levinson algorithm
}
\usage{Durbin_Levinson_rho(rho)}

\arguments{
  \item{rho}{a vector of autocorrelations}
}

\value{a vector of partial autocorrelations.}

\references{
Shumway, R. H., & Stoffer, D. S. (2006). Time series analysis and its applications with r examples (2nd ed.). New York: Springer.  (p.113-114)
}
\author{Zijun Ke <keziyun@mail.sysu.edu.cn> and Zhiyong Zhang <zhiyongzhang@nd.edu>}
