\name{pautocorr-package}
\alias{pautocorr-package}
\alias{pautocorr}
\docType{package}
\title{
Asymptotic methods and resampling techniques for testing (partial) autocorrelations
}
\description{
Two asymptotic methods (the 1/T approximation method and the Bartlett's formula method) and two resampling techniques (the surrogate data method and the vectorized moving block bootstrap) are implemented to test the significance of (partial) autocorrelations at certain lags.
}
\details{
\tabular{ll}{
Package: \tab autocorr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-25\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
Bartlett_Shumway

BartlettSE

daily.moral

Durbin_Levinson_rho

BCa.CI

corr.VMB

get.ahat

get.T

JN.NB

JN.VMB

pautocorr.test

pair.VMB

Perc.CI

plot.pautocorr

Surrogate

Test.CI

VMB
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
Zijun Ke and Zhiyong Zhang
Maintainer: Zijun Ke <keziyun@mail.sysu.edu.cn>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
